fix bug in predicate pushdown
Very simple predicates were not
pushed down to the scan level

Fixed and added proper tests.
Will add more.
